Make load_alerts a shared_task
